BHA-193 |Refactoring - added test for sim executors and BatchNeo4j |Gaurav
* fix mocking for GraphProvider and AplicationConfig

+++ src/main/scala/com/bharatsim/engine/execution/control/DistributeStateControl.scala
@@ -35,7 +35,7 @@ class DistributeStateControl(context: Context) {
         )
         .onComplete {
           case x: scala.util.Success[Record] =>
-            val nodeId = x.value.get("newStateId").asInt()
+            val nodeId = x.value.get("newStateId").asLong()
 
             state.setId(nodeId)
             state.enterAction(context, statefulAgent)
